requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.20.0
pynvml>=11.0.0
pytest>=7.0.0
pyyaml>=6.0.0
librosa>=0.10.0
soundfile>=0.12.0
tqdm>=4.65.0
# Voice Embedding dependencies
speechbrain==0.5.16
transformers>=4.30.0
# Visualization dependencies
matplotlib>=3.7.0
scikit-learn>=1.0.0
seaborn>=0.12.0
gTTS==2.3.2
pydub==0.25.1
